move all layout components to one file, because there was a error related to Vite 'document is not defined', web workers is not working
