--- src/app/Layouts/components/footer/index.ts
@@ -1,1 +1,0 @@
-export { Footer } from "./footer";
