--- src/app/Layouts/components/header/index.ts
@@ -1,1 +1,0 @@
-export { Header } from "./header";
