--- src/app/Layouts/components/mobile-header/index.ts
@@ -1,1 +1,0 @@
-export { MobileHeader } from './mobile-header';
