--- src/app/Layouts/index.ts
@@ -1,1 +1,0 @@
-export { CurrentLayout } from './current-layout';
