--- src/app/Layouts/layout-types/index.ts
@@ -1,2 +1,0 @@
-export { FullWidthLayout} from './full-width-layout';
-export { MobileLayout } from './mobile-layout';
