--- src/app/Layouts/layout-types/types.ts
@@ -1,5 +1,0 @@
-export interface ILayoutType {
-    headerHeightPx: string;
-    footerHeightPx: string;
-    navWidthPx: string;
-}
