+++ src/app/shared/utils/index.ts
@@ -1,5 +1,4 @@
 export { createOrbitLanes } from './createOrbitLines';
-export { getCurrentLayout } from './get-current-layout';
 export { getLanguage } from './getLanguage';
 export { getPath } from './getPath';
 export { getSatellitePosition } from './getSatellitePosition';
